requirements.txt
click>=8.1
httpx>=0.27
rich>=13.0
python-dotenv>=1.0
